+++ CHANGELOG.md
@@ -9,6 +9,7 @@ Please ADD ALL Changes to the UNRELEASED SECTION and not a specific release
 ### Security
 ### Added
 - Unit tests for FunFair.BuildCheck.Helpers to achieve 100% code coverage
+- Unit tests for FunFair.BuildCheck.Interfaces to achieve 100% code coverage
 ### Fixed
 ### Changed
 ### Deprecated